backend/requirements.txt
asgiref==3.2.10
attrs==20.2.0
category-encoders==2.2.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-extensions==3.0.9
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
djangorestframework==3.11.1
drf-yasg==1.17.1
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
lightgbm==3.0.0
MarkupSafe==1.1.1
numpy==1.19.3
oauthlib==3.1.0
packaging==20.4
pandas==1.1.4
patsy==0.5.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
statsmodels==0.12.1
swagger-spec-validator==2.7.3
threadpoolctl==2.1.0
uritemplate==3.0.1
urllib3==1.25.11
xgboost==1.2.1
zipp==3.4.0
